# docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Uptime Cairn API
  version: 1.0.0
  summary: All-in-one open source uptime monitoring and reporting platform.
  description: |
    The complete public surface of Uptime Cairn. The dashboard is the first
    client of this API, never a privileged one — there are no internal
    endpoints the UI can reach that an API key cannot.

    ## Authentication

    Two schemes, either of which authenticates any non-public operation:

    - `ApiKeyAuth` — `Authorization: Bearer cairn_<key>`. Scoped, expiring,
      revocable. The scheme for automation, CI/CD, and Terraform.
    - `SessionCookie` — the browser session the dashboard uses after
      `POST /api/v1/auth/login`. Cookie-authenticated state-changing requests
      additionally require a CSRF token in `X-Cairn-CSRF-Token`.

    Operations tagged `Public` require no authentication: the status page read
    path, subscriber confirmation, and push-monitor ingest.

    ## Pagination

    Every list endpoint paginates by opaque cursor, keyed on `(updated_at, id)`,
    per ADR-004 — uniformly, with no exception for installs small enough that
    sending everything would "work". Offsets are unstable against rows that
    change while a user is paginating, and status changes constantly.

    Cursor responses carry no total count. A count for the current filter comes
    from `GET /api/v1/monitors/membership`, which a client polling for live-view
    freshness is already calling.

...

    ## Live updates

...

...

...

...

...

...
servers:
...
security:
  - ApiKeyAuth: []
  - SessionCookie: []
...
paths:
...
  /api/v1/system/info:
    get:
      tags: [System]
      operationId: getSystemInfo
      summary: Instance metadata
      description: Version, build, deployment mode, and enabled capabilities.
      x-cairn-phase: 1
      responses:
        "200":
          description: Instance metadata.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SystemInfo" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/overview:
    get:
      tags: [System]
      operationId: getOverview
      summary: Dashboard summary counts
      description: |
        Aggregate counts for the dashboard header — status tallies, active incidents,
        active maintenance, upcoming expiries. Exists so the dashboard can render a
        summary without fetching every monitor, which is the failure mode this
        project is built to avoid.
      x-cairn-phase: 1
      responses:
        "200":
          description: Summary counts.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Overview" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/monitors:
    get:
      tags: [Monitors]
      operationId: listMonitors
      summary: List monitors
      description: |
        Server-side paginated, filtered, sorted, and searched. This endpoint must remain
        responsive with 5,000 monitors configured — it is covered by the CI load-test gate.

        Filters combine with AND. Repeated array parameters combine with OR within the
        parameter, so `tag_id=a&tag_id=b` matches monitors carrying either tag.
      x-cairn-phase: 1
      x-cairn-scopes: [monitors:read]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/CollectionLimit"
        - $ref: "#/components/parameters/Search"
        - name: sort
          in: query
          description: |
            Result ordering. A keyset cursor can only paginate the ordering it is keyed
            on, so this selects the cursor key as well as the sort — ADR-009 generalises
            it from ADR-004's fixed `(updated_at, id)` to `(sort_field, id)`.

            Sortable fields are stable, stored, non-null columns, and the set is closed.
            **`status` and uptime are deliberately absent.** Ordering by a value that
            changes while you paginate makes a row jump position, so the reader sees it
            twice or never — the exact defect keyset pagination exists to remove. Uptime
            is computed at read time and has no column to key a cursor on. Filter by
            status instead; it answers the question sorting by it was reaching for.

            Name ordering is case-insensitive and identical across storage backends.

            **A `cursor` must be used with the `sort` it was issued for.** Changing sort
            mid-pagination is a `400`, not a silent reset to the first page.
          schema:
            type: string
            default: -updated_at
            enum: [updated_at, -updated_at, name, -name, created_at, -created_at]
        - name: status
          in: query
          description: Filter by current status.
          schema:
            type: array
            items: { $ref: "#/components/schemas/MonitorStatus" }
          style: form
          explode: true
        - name: type
          in: query
          description: Filter by monitor type.
          schema:
            type: array
            items: { $ref: "#/components/schemas/MonitorType" }
          style: form
          explode: true
        - name: tag_id
          in: query
          schema:
            type: array
            items: { type: string, format: uuid }
          style: form
          explode: true
        - name: group_id
          in: query
          schema:
            type: array
            items: { type: string, format: uuid }
          style: form
          explode: true
        - name: enabled
          in: query
          schema: { type: boolean }
        - name: include
          in: query
          description: |
            Optionally embed related data. `last_heartbeat` and `uptime` cost extra
            per-row work; the dashboard list view requests them, exports do not.

            `heartbeats` embeds a bounded run of recent checks per monitor — the
            strip a list view draws under each row. It is resolved for the whole
            page in one statement rather than one request per row, which is the
            fan-out this parameter exists to prevent: a client that requests it
            per row has reintroduced exactly the cost the 5,000-monitor gate
            measures. Its length is `heartbeats_limit`, and it is bounded on
            purpose so the embed's cost tracks the viewport rather than the
            install (ADR-004).
          schema:
            type: array
            items: { type: string, enum: [last_heartbeat, heartbeats, uptime, tags, group] }
          style: form
          explode: false
        - name: heartbeats_limit
          in: query
          description: |
            How many recent checks `include=heartbeats` returns per monitor.
            Clamped to the maximum rather than refused, the same as `limit`.
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        "200":
          description: A page of monitors.
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Monitor" }
                  pagination: { $ref: "#/components/schemas/CursorPagination" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Monitors]
      operationId: createMonitor
      summary: Create a monitor
      description: |
        The monitor begins checking immediately unless `enabled` is false. Its first
        heartbeat is recorded on the next scheduler tick, not synchronously with this call.
      x-cairn-phase: 1
      x-cairn-scopes: [monitors:write]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MonitorWrite" }
      responses:
        "201":
          description: Monitor created.
          headers:
            Location:
              description: URL of the created monitor.
              schema: { type: string, format: uri-reference }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Monitor" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: cairn_<key>
      description: |
        A scoped API key: `Authorization: Bearer cairn_<key>`. Keys carry an explicit
        scope list, an optional expiry, and a last-used timestamp, and can be revoked
        at any time. Every operation names the scopes it requires in `x-cairn-scopes`.
    SessionCookie:
      type: apiKey
      in: cookie
      name: cairn_session
      description: |
        The browser session established by `POST /api/v1/auth/login`. The cookie is
        `HttpOnly`, `SameSite=Lax`, and `Secure` when served over HTTPS.
        State-changing requests authenticated this way must also carry a matching
        `X-Cairn-CSRF-Token` header.
  parameters:
    CollectionLimit:
      name: limit
      in: query
      description: |
        Maximum items to return. The server caps this; requesting more is clamped,
        not rejected.
      schema: { type: integer, minimum: 1, maximum: 200, default: 25 }
    Search:
      name: q
      in: query
      description: Free-text search across the resource's human-readable fields.
      schema: { type: string, maxLength: 200 }
    Cursor:
      name: cursor
      in: query
      description: |
        Opaque cursor from a previous response's `pagination.next_cursor`. Cursors are
        not decodable, not sortable, and not guaranteed stable across releases.
      schema: { type: string }
...
  responses:
...
    Unauthorized:
      description: No credentials were supplied, or they were not valid.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    SystemInfo:
      type: object
      required: [version, mode]
      properties:
        version:
          type: string
          description: Semantic version of the running build.
          examples: ["0.1.0"]
        commit: { type: string }
        built_at: { type: string, format: date-time }
        mode:
          type: string
          enum: [solo, control_plane, probe]
          description: |
            Deployment shape. Solo is the single binary with an embedded probe and
            SQLite; the others are the scaled-mode split from ADR-001.
        storage_engine: { type: string, enum: [sqlite, postgres_timescale] }
        api_version: { type: string, examples: ["v1"] }
        monitor_types:
          type: array
          description: Monitor types this build supports, so clients need not hardcode the list.
          items: { $ref: "#/components/schemas/MonitorType" }
        notification_channel_types:
          type: array
          items: { $ref: "#/components/schemas/NotificationChannelType" }
        capabilities:
          type: object
          description: |
            Feature flags reflecting what this build and configuration can do — the
            mechanism progressive disclosure runs on. The dashboard hides surfaces the
            instance does not have rather than showing dead controls.
          additionalProperties: { type: boolean }
    Overview:
      type: object
      description: Aggregate dashboard counts.
      required: [monitors, generated_at]
      properties:
        monitors:
          type: object
          required: [total, up, down, pending, paused, maintenance]
          properties:
            total: { type: integer }
            up: { type: integer }
            down: { type: integer }
            pending: { type: integer }
            paused: { type: integer }
            maintenance: { type: integer }
        active_incidents: { type: integer }
        active_maintenance_windows: { type: integer }
        certificates_expiring_soon:
          type: integer
          description: Monitors whose observed certificate falls inside its configured threshold.
        domains_expiring_soon: { type: integer }
        generated_at: { type: string, format: date-time }

    # ------------------------------------------------------------------------
    # Setup, sessions, and API keys
    # ------------------------------------------------------------------------
...